move app_url const
