+++ lib/consts.tsx
@@ -53,5 +53,3 @@ export const PULSATING_COLORS = {
   memories: '#45b2c7',
   plannet: '#70c8f7',
 }
-
-export const APP_URL = 'https://maravilla.city'
