+++ lib/farcaster/consts.ts
@@ -1,0 +1,1 @@
+export const APP_URL = 'https://maravilla.city'
